refactor: colocalize standardization with adapt processes
