+++ src/config/index.ts
@@ -1,0 +1,3 @@
+export * from "./external/index.js";
+export * from "./standard/index.js";
+export * from "./manage.js";
